Fixed location & refactor rest
